+++ src/App.js
@@ -1,5 +1,4 @@
-import Login from 'sections/auth/Login';
-import Signup from 'sections/auth/Signup';
+
 import Router from 'shared/Router';
 import { GlobalStyles } from 'styles/GlobalStyles';
